requirements.txt
fastapi
uvicorn[standard]
gunicorn
sqlalchemy
alembic
pydantic
python-jose[cryptography]
passlib
python-multipart
psycopg2-binary
python-dotenv
tensorflow
pillow
opencv-python-headless
numpy
